+++ ProductCatalog.Service/Interfaces/IProductService.cs
@@ -6,8 +6,8 @@ namespace ProductCatalog.Service.Interfaces
     {
         Task<(IEnumerable<Product> Items, int TotalItems)> GetAllProductsAsync(
             int? categoryId,
-            double? minPrice,
-            double? maxPrice,
+            decimal? minPrice,
+            decimal? maxPrice,
             bool? isActive,
             bool? inStock,
             string? sortBy,
